% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{bin_tracking_data}
\alias{bin_tracking_data}
\title{Bin tracking data}
\usage{
bin_tracking_data(data, spatial = 1000, temporal = "week")
}
\arguments{
\item{data}{Tracking data to be aggregated with the columns
\code{"organismID"}, \code{"decimalLongitude"}, \code{"decimalLatitude"},
\code{"eventDate"} and optionally \code{"detections"}.}

\item{spatial}{Spatial resolution in meters to be used to aggregate the data
(default \code{1000}).}

\item{temporal}{Temporal resolution to be used to aggregate the data (default
\code{"week"}).}
}
\value{
Data frame with columns \code{"organismID"}, \code{"decimalLongitude"},
  \code{"decimalLatitude"}, \code{"eventDateArrival"}, \code{"eventDateDeparture"}
  and \code{"detections"}.
}
\description{
Aggregate tracking data with a specific spatial and temporal resolution
}
\details{
The aggregation is done in a spatiotemporal grid based on the
  provided spatial and temporal resolution. The spatial aggregation is done
  on a Behrmann equal area grid with as cell size the specified spatial
  resolution in meters. The temporal aggregation accepts either on of the
  following string literals: \code{"day"}, \code{"week"}, \code{"month"},
  \code{"year"} or a date time format string e.g. \code{"\%Y-\%m-\%d"}
}
